Upgrade by Antigravity code agent
