+++ .gitignore
@@ -128,3 +128,4 @@ yarn-error.log
 .pnp.js
 # Yarn Integrity file
 .yarn-integrity
+.vercel
