+++ package.json
@@ -36,14 +36,14 @@
     "@mdx-js/react": "^2.2.1",
     "@popperjs/core": "^2.11.6",
     "bootstrap": "^5.2.3",
+    "decap-cms-app": "3.6.4",
     "eslint": "^8.30.0",
-    "esm": "^3.2.25",
     "gatsby": "^5.3.2",
     "gatsby-plugin-alias-imports": "^1.0.5",
+    "gatsby-plugin-decap-cms": "^4.0.0",
     "gatsby-plugin-graphql-loader": "^1.0.2",
     "gatsby-plugin-image": "^3.3.2",
     "gatsby-plugin-mdx": "^5.3.1",
-    "gatsby-plugin-netlify-cms": "^7.3.0",
     "gatsby-plugin-react-helmet": "^6.3.0",
     "gatsby-plugin-sass": "^6.3.1",
     "gatsby-plugin-sharp": "^5.3.2",
@@ -54,9 +54,7 @@
     "gatsby-transformer-yaml": "^5.3.0",
     "graphql": "^16.6.0",
     "graphql-tag": "^2.12.6",
-    "jquery": "^3.6.2",
-    "netlify-cms-app": "^2.15.72",
-    "node-sass": "^8.0.0",
+    "jquery": "^3.7.1",
     "pug-as-jsx-loader": "^1.0.88",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -70,8 +68,8 @@
     "@babel/core": "^7.20.5",
     "@babel/node": "^7.20.5",
     "@types/node": "^18.11.10",
-    "@types/react": "^18.0.26",
-    "@types/react-dom": "^18.0.9",
+    "@types/react": "^18.2.0",
+    "@types/react-dom": "^18.2.0",
     "babel-loader": "^9.1.0",
     "webpack": "^5.75.0"
   }
