+++ src/css/theme.scss
@@ -398,3 +398,7 @@ img {
     font-size: 11px;
   }
 }
+
+.email {
+    white-space: pre-wrap;
+}
